requirements.txt
discord.py==2.3.2
aiofiles==23.2.1
aiohttp==3.9.1
googletrans==3.1.0a0
requests==2.31.0
plyer==2.1.0
colorama==0.4.6
pystyle==2.9
Pillow==10.1.0
tls-client==0.2.2
spotipy==2.23.0
pypresence==4.3.0
pywin32==306
pyautogui==0.9.54
